+++ hipamd/include/hip/amd_detail/amd_hip_cooperative_groups.h
@@ -96,6 +96,8 @@ class thread_group {
   __CG_QUALIFIER__ unsigned int cg_type() const { return _type; }
   //! Rank of the calling thread within [0, \link num_threads() num_threads() \endlink).
   __CG_QUALIFIER__ __hip_uint32_t thread_rank() const;
+  //! Rank of the block in calling thread within [0, \link num_threads() num_threads() \endlink).
+  __CG_QUALIFIER__ __hip_uint32_t block_rank() const;
   //! Returns true if the group has not violated any API constraints.
   __CG_QUALIFIER__ bool is_valid() const;
 
@@ -203,6 +205,8 @@ class grid_group : public thread_group {
  public:
   //! @copydoc thread_group::thread_rank
   __CG_QUALIFIER__ __hip_uint32_t thread_rank() const { return internal::grid::thread_rank(); }
+  //! @copydoc thread_group::block_rank
+  __CG_QUALIFIER__ __hip_uint32_t block_rank() const { return internal::grid::block_rank(); }
   //! @copydoc thread_group::is_valid
   __CG_QUALIFIER__ bool is_valid() const { return internal::grid::is_valid(); }
   //! @copydoc thread_group::sync
@@ -274,6 +278,10 @@ class thread_block : public thread_group {
   //! @copydoc thread_group::thread_rank
   __CG_STATIC_QUALIFIER__ __hip_uint32_t thread_rank() {
     return internal::workgroup::thread_rank();
+  }
+  //! @copydoc thread_group::block_rank
+  __CG_STATIC_QUALIFIER__ __hip_uint32_t block_rank() {
+    return internal::workgroup::block_rank();
   }
   //! @copydoc thread_group::num_threads
   __CG_STATIC_QUALIFIER__ __hip_uint32_t num_threads() {
@@ -353,7 +361,6 @@ class tiled_group : public thread_group {
   __CG_QUALIFIER__ unsigned int thread_rank() const {
     return (internal::workgroup::thread_rank() & (coalesced_info.tiled_info.num_threads - 1));
   }
-
   //! @copydoc thread_group::sync
   __CG_QUALIFIER__ void sync() const { internal::tiled_group::sync(); }
 };